# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic Ecommerce Mission Control API
  version: 9.0.0
  description: |
    Public v1.x.y contract for the Agentic Ecommerce Go backend consumed by
    agentic-ecommerce-web. This spec intentionally excludes private
    mobile-only operator endpoints and credential-bearing surfaces.
    Protected endpoints use short-lived HMAC-signed JWT bearer tokens and RBAC roles:
    admin, operator, and viewer.

    Version policy: v1 endpoints are stable through host v9.x. The
    v2 preview surface (api/openapi-v2-preview.yaml) is opt-in only
    and subject to change without notice. See docs/api-versioning.md
    for negotiation, deprecation timeline, and Accept-header opt-in.
  x-api-version-policy: |
    v1 endpoints are stable through host v9.x; v2 preview endpoints are
    subject to change without notice. See docs/api-versioning.md.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/products:
    get:
      summary: List storefront products
      operationId: listProducts
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated product list
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
              description: Present when the request origin matches ECOMMERCE_ALLOWED_ORIGIN.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '401':
          description: Missing or invalid JWT bearer token when auth is configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Origin rejected by CORS policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a product
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid JSON body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate product (SKU or slug collision).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed (missing SKU, invalid price, etc).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    options:
      summary: Products CORS preflight
      operationId: productsPreflight
      responses:
        '204':
          description: Preflight accepted.
        '403':
          description: Origin rejected by CORS policy.
...
  /api/v1/sync/dlq:
    get:
      summary: List marketplace sync DLQ records
      operationId: listMarketplaceDLQ
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Marketplace sync DLQ records currently visible to the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketplaceDLQListResponse'
  /api/v1/sync/dlq/{id}/replay:
    post:
      summary: Replay a marketplace sync DLQ record
      operationId: replayMarketplaceDLQ
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '202':
          description: Marketplace replay workflow accepted by Temporal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStartResponse'
        '400':
          description: Invalid DLQ record ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: DLQ record was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal rejected the replay workflow start request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/sync/conflicts:
    get:
      summary: List pending WooCommerce sync conflicts
      operationId: listSyncConflicts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending conflicts queued for manual review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictListResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HMAC-signed JWT access token issued by `/api/v1/auth/login` or `/api/v1/auth/refresh`.
  schemas:
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: integer
          minimum: 0
          description: Amount in the smallest currency unit (for example, cents).
        currency:
          type: string
          enum: [AUD, USD, GBP, EUR]
    Product:
      type: object
      required: [id, sku, title, slug, price, stock, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        slug:
          type: string
          pattern: '^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$'
        price:
          $ref: '#/components/schemas/Money'
        stock:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [draft, active, archived]
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProductRequest:
      type: object
      required: [sku, title, price]
      properties:
        sku:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        slug:
          type: string
          description: Auto-generated from title if omitted.
        description:
          type: string
        price:
          $ref: '#/components/schemas/Money'
        stock:
          type: integer
          minimum: 0
          default: 0
        status:
          type: string
          enum: [draft, active, archived]
          default: draft
    ProductListResponse:
      type: object
      required: [products, total, page, per_page]
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
...
    WorkflowStartResponse:
      type: object
      required: [workflow_id, run_id, status, task_queue]
      properties:
        workflow_id:
          type: string
        run_id:
          type: string
        status:
          type: string
          enum: [started]
        task_queue:
          type: string
          enum: [ec-workflows]
    MarketplaceSyncEvent:
      type: object
      required: [tenant_id, provider, entity_type, entity_id, operation, version]
      properties:
        tenant_id:
          type: string
        provider:
          type: string
        entity_type:
          type: string
          enum: [product]
        entity_id:
          type: string
        external_id:
          type: string
        operation:
          type: string
          enum: [upsert, delete]
        version:
          type: string
        payload:
          type: object
          additionalProperties: true
    MarketplaceDLQRecord:
      type: object
      required: [id, event, attempts, reason]
      properties:
        id:
          type: string
        event:
          $ref: '#/components/schemas/MarketplaceSyncEvent'
        attempts:
          type: integer
          minimum: 1
        reason:
          type: string
...
    MarketplaceDLQListResponse:
      type: object
      required: [records, total]
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/MarketplaceDLQRecord'
        total:
          type: integer
          minimum: 0
    SyncConflictField:
      type: object
      required: [field, local_value, remote_value]
      properties:
        field:
          type: string
          enum: [title, price, stock, description]
        local_value:
          type: string
        remote_value:
          type: string
    SyncConflict:
      type: object
      required: [id, sku, remote_id, status, fields, created_at]
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        sku:
          type: string
        remote_id:
          type: integer
        status:
          type: string
          enum: [pending, resolved]
        fields:
          type: array
          items:
            $ref: '#/components/schemas/SyncConflictField'
        resolution:
          type: string
        note:
          type: string
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
    ConflictListResponse:
      type: object
      required: [conflicts]
      properties:
        conflicts:
          type: array
          items:
            $ref: '#/components/schemas/SyncConflict'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...